Get experimental roll controller working with only top lateral face as reference!

+++ Shared Projects/Flight/STU Flight Controller/Controllers/STUOrientationController.cs
@@ -51,9 +51,9 @@ namespace IngameScript {
                     return false;
                 }
 
-                public void SetRoll(float rollSpeed) {
+                public void SetRoll(double rollSpeed) {
                     foreach (var gyro in Gyros) {
-                        gyro.Roll = rollSpeed;
+                        gyro.Roll = (float)rollSpeed;
                     }
                 }
